+++ modules/ACE/ace/Shared_Memory_SV.h
@@ -46,14 +46,14 @@ public:
   // = Initialization and termination methods.
   ACE_Shared_Memory_SV (void);
   ACE_Shared_Memory_SV (key_t id,
-                        int length,
+                        size_t length,
                         int create = ACE_Shared_Memory_SV::ACE_OPEN,
                         int perms = ACE_DEFAULT_FILE_PERMS,
                         void *addr = 0,
                         int flags = 0);
 
   int open (key_t id,
-            int length,
+            size_t length,
             int create = ACE_Shared_Memory_SV::ACE_OPEN,
             int perms = ACE_DEFAULT_FILE_PERMS,
             void *addr = 0,
@@ -66,14 +66,14 @@ public:
   virtual int remove (void);
 
   // = Allocation and deallocation methods.
-  /// Create a new chuck of memory containing <size> bytes.
+  /// Create a new chuck of memory containing @a size bytes.
   virtual void *malloc (size_t = 0);
 
   /// Free a chuck of memory allocated by <ACE_Shared_Memory_SV::malloc>.
   virtual int free (void *p);
 
   /// Return the size of the shared memory segment.
-  virtual int get_segment_size (void) const;
+  virtual size_t get_segment_size (void) const;
 
   /// Return the ID of the shared memory segment (i.e., a System V
   /// shared memory internal id).
